moving dummy classes to test
